refactor : api 연동
