+++ src/components/profile/SetGoalModal.tsx
@@ -3,7 +3,7 @@ import { MdCancel } from 'react-icons/md';
 import { goalData } from '@/data/goalData';
 import Image from 'next/image';
 import { useState } from 'react';
-import { setGoal } from '@/api/setGoal';
+import { setGoal } from '@/api/profile';
 import toast from 'react-hot-toast';
 
 export default function SetGoalModal({ onClose }: { onClose: () => void }) {
